requirements.txt
# requirements.txt - 依赖包

# 阿里云百炼 SDK
dashscope>=1.17.0

# Word文档读取
python-docx>=0.8.11

# PDF文档读取
PyPDF2>=3.0.0

# 网页内容提取
beautifulsoup4>=4.12.0
requests>=2.31.0

# 音频处理
pydub>=0.25.1

# Edge TTS（微软免费语音合成）
edge-tts>=6.1.0

# WebSocket客户端
websocket-client>=1.8.0

# FastAPI
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
